No more hashmaps :(

+++ 5/part1_old.h
@@ -1,7 +1,7 @@
 typedef struct LinkedNode LinkedNode;
 struct LinkedNode{
-    int key;
-    int val;
+    long key;
+    long val;
     LinkedNode* next;
 };
 
@@ -10,11 +10,11 @@ typedef struct {
     int capacity;
     int size;
 } HashTable;
-int hash(int x);
-int hashIndexOf(int h, int length);
+long hash(long x);
+int hashIndexOf(long h, int length);
 HashTable* initHashTable(int capacity);
 void resizeHashTable(HashTable** htp);
 void destroyLinkedList(LinkedNode* head);
-void hashPut(HashTable** htp, int key, int val);
+void hashPut(HashTable** htp, long key, long val);
 void destroyHashTable(HashTable* ht);
-int hashGet(HashTable* ht, int key);
+long hashGet(HashTable* ht, long key);
